add build command and refactor architecture
Add a build command. If no additional parameter is passed it builds
either all apps in the repository or if it's run inside an application
directory the current application.
It also accepts "all", the path to an application directory or an
application name to indicate what should be build.

A build is done by running the command specified build_command parameter
in the application or repository config file.
The parameter in the application config file overwrites the one in the
repository config.

To be able to integrate the build command smoothly, some parts were
refactored:
- move discover functionality to Repository and App struct
- remove command/ctx.go