cfg: add baur_version field to repository config
Add a "baur_version" field to the repository config.
It contains the used version of baur.
If the the config doesn't have the value, the current version is written
to it when a baur command is called the first time.

Having the version number in the repository config will allow us to
notify the user that:
- his baur version is too old for a repository config. This is likely to
  happen when baur configs were updated in a git repository but some
  users haven't updated their local baur version.
- inform the user that he has to update manually his config files on
  some breaking changes