cfg: move converting application dirs to absolute paths to baur
When a repository config was read, cfg.RepositoryFromFile() was
converting all relative paths in the Discover.application_dirs field to
absolute paths before
Move this logic to baur.NewRepository(). The cfg package should only
read, write and validate the configs. It should not modify the values.

The baur package transforms them if needed.

This also fixes that if a repository config was rewritten by the
cfg.package, the values of Discover.application_dirs were converted to
absolute paths.