cfg: add tests that validate an example conf read from a file
Validating an Example CFG might be succesful if we the value returned
from ExampleRepository() and ExampleApp() but might fail after the same
object was serialized to toml, written to a file and then read.
This can happen if the toml serialization works different then expected.
Add tests for those cases.