baur: exit with an error if baur_version does not match baur version
Exit with an error if the baur_version string in the baur.toml file has
a different major or minor version then the baur version that is run.
This should ensure that users run an uptodate baur version.

In the future when we change to major version 1, configs should stay
compatible between the same major version. The check will be reworked.