postgres: store remote destinations of artifacts in seperate table
Normalize the database further, store upload informations in a seperate
table instead of in the artifact table.
Because upload and artifact informations were decoupled, multiple builds
can now also link to the same artifact record.

The commit also outsources the logic to create a record if it doesn't
exist to a seperate function, to reduce code duplication.