artifact: add Digest() functions
For file artifacts it will return the SHA256 digest of the file.

For docker containers it's not trivial to get a digest for all it's
content:
- the digest shown by "docker images --digests", is the digest of the
  image manifest that is created from the docker registry
- docker image layers contain timestamps of files, also when an image is
  created with the same content, the layer digests differ because of the
  timestamps

We are using the image id for now, it's at least always the same for the
same image independent if it was pushed/pulled to/from a registry.