storage: record artifact size
Record the size of artifacts in the storage layer.
The Artifact interface got a new method Size() that returns the size of
an artifact in Byte.

To retrieve the size of a docker container, we need a docker client to
request the size.
An option would be to pass a docker.Client in the constructor for a
DockerArtifact but thats kind of ugly and we would need to pass the
client through the repository, to the NewApp call and further.

For now the Size() method takes as first Argument a ArtifactBackends
struct which contains references to artifact backends.
This should be improved.
Maybe it's better to introduce a runtimeConfiguration object or simply
store required runtime references (like the docker client) in the baur
package and access it when needed.