baur: support specifying source files of applications in .app.toml files
The commit adds a [SourceFiles] section to the app.toml. It contains a
parameter `paths` to specify multiple glob paths to source files of the
application.

In a following commit, baur will keep track of the checksums of those
source files. Those informations will be used to decide if an
application needs to be rebuild or if a build already existed with the
same source files.