build: store the sources of an build in the database
Store the source in the database with the build.
The implementation is not ideal, currently first all digests of all
applications are calculated and then the builds are started.
It's probably better to couple calculating the digests to the build
steps, so it can be done in parallel for multiple apps.
Or have some async mechanism like for builds and uploads