baur: support specifying docker images as build sources/input
Add a [[DockerSource]] section to the config file to allow specifying
docker images as build source.

They need to be specified by digest instead of per tag to ensure the
reference is unambiguous.

DockerSources are handled smiler then FileSource and GitFileSources.
They are fetched from the cfg and then initiated as DockerSource
implementing the SrcResolver interface to keep the usage consistent.
Because they are unambiguous they actually don't need to be resolved and
Resolve() always returns only a single RemoteDockerImg object. It's only
to keep the interface for now.

The BuildInput interface was introduced to be able to handle File and
Docker source inputs the same way.
Sourcefiles are not stored by their plan relative path anymore in the
db, they are now stored as URL.
Currently still with an relative path to the application dir, this will
change to a repository relative path.