src: calculate total src digest and store it in the database
Calculate a digest over all digests of all sources.
This is done by sorting the digests by algorithm and numerical by their
hash, concatenating those to a string and creating a SHA384 digest of it.
To be able to compare the digests sums numerical, they are now stored as
bigint in the Digest.

The digest is stored in the database.

The commit also fixes some minor debug output issues