postgresql: use upsert with ON CONFLICT instead of rollback
To only insert records in some tables if they don't exist switch to
using an INSERT with ON CONFLICT update statement.
It's faster then trying to do an INSERT and if it fails doing a rollback
and SELECT.

The Upsert approach has the disadvantage that the serial sequence is
incremented for every run UPDATE.
This should not be a problem. The serial type is big enough for doing
250k updates every day for 23years.