inputs: include path of BuildInput files in digest
The digest of BuildInput files was computed only over the content of the
file.
This had the problem that if a file was moved but it's content didn't
change an application wasn't rebuild because It's TotalInputDigest did
not change.
This causes a problem in the following scenario:
- BuildInput A contains the path to BuildInput B,
  the path is correct but BuildInput B is a different path in the
  repository
- BuildInput B is moved to the path in seperate commit

Until now the application would not have been rebuild.
This is fixed by including the repository relative path of the file in
the hash.

To be able to add the path of the file to the digest, the sha384 package
was refactored.
It now offers a Hash object to which data can be added and the digest is
retrieved by a separate function call. This allows to add the content of
the file and in a septate call the path of the file to the hash and then
retrieving it's Digest.