view: refactor package
The logic to format output is moved into a format package which is only
responsible to print Rows in a formatted manner.
Data can be written Row by Row to the Formatter instead of only all
output in one call, this will allow to have streaming output for
commands that take a longer time.
The dataprovider package was removed. The caller directly writes to the
Formatter interface.
This allows the caller to have full control of the format of each value.
It also makes it more intuitive too use.

Previously the dataprovider defined the format and the order of the data
that is passed to the formatter. To change the value for one value when
a different formatter was used, a new dataprovider needed to be added
with duplicate code.