command: refactor sort flag
The sort flag is generalized to handle arbitrary field values.
Previously the supported field values were hardcoded in the flag
implementation.
Now a mapping of commandline parameters to storage.Fields that the flag
should support be passed in the constructor.

The flag now also creates a usage output string via the Usage() method.
String() is changed to return an empty string, cobra calls the function
to show the default value in the usage output.
It was shown as: "(default FieldUndefined-invalid)".
When String() returns an empty string the whole default part is omitted.