command: do not write a header in CSV format
Do not write a header when printing something in csv format.
The CSV RFC does not specify a way to distinguish a header from fields.

The user would have to skip always the first line when processing the
CSV output manually. This makes it more cumbersome to use.

We might want to add later a parameter to print csv headers
"--with-csv-header"