build: only use docker auth data from env variables if set
The values of the env variables DOCKER_USERNAME, DOCKER_PASSWORD were
always used to authenticating at docker.
It was only possible to authenticate at the docker registry by setting
the env variable.

When they are empty, do not pass any authentication data to the docker
client.
This way the docker library will user authentication data from the
docker credentials helpers in this case.