config: set reasonable default BuildInputs in example app configs
When an .app.toml file is generated via "baur apps init", it now
contains uncommented BuildInputs lines to include the .app.toml file and
the all files in the app directory.

The .app.toml file should be included as build input because if e.g. the
build_command, the upload destination, or the buildinputs change the
applications should be rebuild to ensure the changes have an effect.