cfg: remove [Build.command] from repository config
Setting a default build command in the repository configuration can be
handy but has the issue, that all applications need to add the
.baur.toml file as buildinput.
Otherwise they are not able to detect that the build command changed in
the repository wide configuration.
Adding the .baur.toml file as buildinput isn't feasible, it means that
all applications are rebuild when e.g. the postgres url changes.

- Remove the Build.Command parameter from the repository config
- Setting a build.command in the .app.toml file is now mandatory
- Change the build.command to "make" and uncomment the parameter in the
  example configuration file