apps ls: show git-commit, switch order of build-id and build-status
Print the build status before the build id, like in the previous baur
versions. It's more straightforward, the following fields all relate to
the build status.
Because the build-status is not the last field anymore the formatting
broke because of the color escape sequences. Tabwriter can't cope with
those, remove the coloring.

Add a git-commit field to the output, it contains the git-commit of an
existing build.