fix: error messages got lost because in some cases
When running baur in a CI job, error messages sometimes got lost.
It was probably because stderr was not flushed in time when baur existed
in the CI environment.

The previous log implementation was not appending a newline to each
message to ensure it's flushed.
This commit switches the log package to be based on the Golang log which
takes care about it.
The S3, upload and exec package do not import the log package anymore
and offer functions to set a logger instead.

Non-error and debug messages are written to STDOUT via the fmt package.