cfg: replace baur_version with config_version in repositoy config
The repository config contained the baur version with that the config
file was created.
If baur was updated and the version number in the config file was not
the same, baur was showing an error.
This had the problem that the configuration syntax in the config file is
not linked to the baur version, the major version might change but the
config file from the previous version can still be compatible, with the
current solution you would nonetheless get an error and be forced to
update the parameter.

baur tool and config version are separated and have to be incremented
independently.
The config_version is a simple integer value, that is incremented on
every non-compatible config file syntax change.
This also allows to do the validation check in the cfg package.