storage: rename upload.url, input.url field to uri
URI fields allow more freedom, we don't have to pass a scheme to be
valid.

Passing a reasonable scheme in URLs was impossible for docker images.
Docker doesn't have an own scheme, registries are using https:// but
docker does not support a scheme in paths passed to docker commands.

Changing it to URI will allow to use docker paths that can be copy &
pasted to docker commands