cfg: move check if postgresql uri is available
Do not check in the cfg package if the postgresql_url parameter is set
anymore.
Instead check before establishing the postgresql connection if the uri
is either set in the config or via the environment variable.

This allows to only set the env variable and keep the configuration
empty in the config