docker: switch vendor repository, support auth via config.json
docker.io/go-docker writes on the github page that it provides the
official docker client lib but it wasn't updated since more then a year.

Switch to using https://github.com/fsouza/go-dockerclient.

The docker package was refactor to support using authentication data
from Docker's config.json files.

If the docker client is via NewClient() without username and password,
it reads the auth data from the config.json.
If the destination push URI that is passed to Upload() contains a
hostname, an entry for this hostname in the config.json is looked up.
If the destination push URI contains no hostname part, we search for an
auth entry for "https://index.docker.io/v1/". If no entry exist, we use
the first found auth entry.