inputs: remove support DockerImageInputs
The option to specify Docker image references as BuildInputs is not
required.

In most cases a build-script or Makefile is called as command which runs
the build step in a Docker Container. The build-script which should be
specified as BuildInput contains the reference to the docker image.
It's not needed to specify the DockerImage as BuildInput again.