upload: add local file copy uploader
Add a new BuildOutput uploader that copies the build artifacts to a
local directory.

This can be useful when baur is run on a CI server and the build
outputs are stored directly in the filesystem, it might be a network
mount.
The section is called "[Build.Output.File.FileCopy]" in the .app.toml
config and has a parameter to set the destination path.

The Uploader creates the director if it does not exist.
If a file already exist with the same path it is overwritten.
The copied file has the same permission bits then the source file.