cfg: set connect_timeout parameter in example PostgreSQL connection uri
Connection timeouts can be set in the postgreSQL connection string.
If no timeout is set baur waits forever if the server is not reachable.
It's better to set a timeout to make baur exit with an error after sometime if
connection establishment failed, add the parameter to the example config that is
created with "baur init repo"