resolve: ignore current environment vars when discovering go source files
When a baur repository was inside the GOPATH and contained a Go Application
using Go-Modules resolving GoSources failed.
It failed because Go-Modules are ignored inside the GOPATH.

A fix is to add "GO111MODULE=on" to Build.Input.GolangSources.environment for
the specific application.

Until now all environment variables of the host were used when resolving go
source files. With this commit only a couple strictly  required environment
variables from the host are inherited.
When a custom GOPATH on the host is used in which the baur repository resides,
resolving gosource files won't fail for this case anymore.
It should also help to reduce other unexpected behaviour that might have
happened because of some env vars set on the host.

Thanks to @joshsteveth for reporting the issue.