resolve: fix: Go stdlib packages added as source
In some circumstances it happened that files from Go's stdlib were discovered as
build input files.
This happened if the packages.Package contained relative paths to the source
file. In this case the strings.HasPrefix() check failed.

Convert the GoSource file paths to absolute paths before checking if they are in
the GOROOT dir.