resolve: fail with error if GOROOT directory does not exist
build.Default.GOROOT only determines the GOROOT path via environment variable.
If the Environment variable is not set, the GOROOT path that was set during
compilation is used.
This can be a different one, then the one in the current host.

Check if the build.Default.GOROOT directory exist, if not fail with an error.
If we discover the wrong GOROOT, we are not able to distinguish stdlib from non
stdlib packages.