resolve: run "go env root" to determine goroot if env var is unset
baur required that one most installation the GOROOT environment variable had to
be set, either on the system or via the .app.toml file.
If this wasn't the case, resolving go source files failed because GOROOT could
not be determined.

If the GOROOT environment variable is not set, run "go env GOROOT" to determine
the GOROOT.
It's the only reliable way to find out the GOROOT path.

The used environment is now also debug logged. This might help to debug problems
in the future.