resolve: improve error message if GitFiles path did not match anything
Extract the paths from the git-ls error message that did not match any files and
return it in an error message.
This enables the user to know exactly which path spec in the .app.toml files
causes the error.

Previously only the message 'file(s) do not exist:' where shown.