resolve: support absolute path for build input files
baur always joined the path of input file with the current app directory.  This
made absolute path specifations invalid, e.g. it was not possible to add a file
in /etc/ as build input.
This can be a valid case if e.g. there is a file that identifies your build
environment.

Only join the build input path with the app directory if it's a relative pathild
input path with the app directory if it's a relative path