docs: update readme
Remove the "go get" instruction from the README, only suggest to install baur
from a release archive. When baur is installed via "go get" the intern version
variable can not be set to the current commit.
When somebody submits later a bugreport it won't be possible to figure out from
which commit his baur version was build which makes debugging much harder.

Also add subsection in Quickstart to make it easier readable and change the
wording slightly.