exec: refactor package
The exec package is refactored completely.
Commands are run by creating a Cmd struct and calling it's Run() method.
The Run() can be configured by calling methods on the Cmd() struct that can be
chained like:
	ShellCommand("ls -1").Directory("/").ExpectSuccess().Run()

Run() now returns an ExitCodeError() if ExpectSuccess() was set for the command.
This simplifies having a consistent error message for all command executions and
callers only have to check for the error instead of also doing an additional
exitcode check.

Packages that were using exec were adapted. Previously all commands were
executed in a shell.
Now only the build command and git ls-files is run in a shell.
git ls-files will be changed to also run without a shell in a following PR.