resolve: do not run "git ls-files" in a shell
"git ls-files" was previously run in a shell.
This caused that the paths that were passed to "git ls-files" were resolved by
the shell to filenames.
When the application directory contained a file that was not tracked in Git and
the path for GitFile BuildInput was a glob that matched this file, resolving the
build inputs failed.

It also has the effect that the paths were matched against the files in the
filesystem instead of against the tracked git files.