cfg: make the [Build] section in the .app.toml file optional
.app.toml files are now also valid if they have no [Build] section defined.
This allows to use baur for discovering apps in a repository without building
them with baur.

"baur ls apps" shows the status "Build Command Undefined" for apps that have no
[Build] section.
When "baur build" is called, apps without a [Build] section are skipped.

The "baur build" output is now more similar with and without "--force" option.
In both cases it now lists the applications that will be build before starting
the build.
If --force is passed, apps that will be build will be shown with Pending status.

The application list that is printed by "baur build" is now also sorted.