cfg: use order of fields in struct when Marshalling to toml
When marshalling a config object to a file, use the order of the fields in the
struct files instead of an alphabetical order.
This will allow to order the fields in a config file created with "baur init"
that makes sense.

To not change the order of the fields in the config in this commit, fields in
the cfg structs were ordered alphabetically.