cfg: add support for include files
This commit introduces a new file format for includes.
An Include file contains multiple Build.Input sections that are identified with
an id. Ids must be unique.
If an id is references in the new input_includes fields in an .app.toml file,
the application includes the input definitions from the include file.

Until now it was common for application that were in the same programming
language that they had the same Build.Input sections in repository.
With this commit the inputs will only need to be defined once instead of having
to duplicate them in multiple .app.toml files.
This makes the input sections easier to maintain.

When NewRepository() is called, the include in the include_dirs are parsed and
added to the Repository struct as a map with their ID as key.
The UnresolvedInputs field of the App struct was changed to a slice of
BuildInputs.
When NewApp() is called, we iterate through the ids in the input_includes fields
and check if they exist in Includes map of the Repository struct.
If they exist, the BuildInput is appended to the UnresolvedInputs of the app.
When inputs are resolved, all BuildInput definition in the UnresolvedInputs
slice are processed.