cfg: support to specify Build.Outputs in include files
Add support to specify Build.Outputs in include files.
The implementation is similar for the Build.Input includes.
The Include cfg struct got an additional slice of BuildInputInclude structs as
member.
The BuildInputInclude struct is the same then BuildInput + an additional id
field.
BuildOutputs can be included in .app.toml files by referencing their id in the
output_includes list.