inputs: include .app.toml and included files automatically
If the .app.toml file or the files that are included are not part of the
BuildInputs of an application, baur is not able to figure out that a rebuild
needs to be done if important settings change like e.g. the build command or the
upload destination.
So far those files had to be added manually to the BuildInput specification.

This became more impractical with the support of including files.
To ensure all include files are tracked, every include file would need to
specify itself as Input file.

With this commit the .app.toml file and files references in the includes setting
are added to the list of BuildInputs automatically by baur.