inputs: add .app.cfg and included files to Build.Inputs
The .app.cfg and included config files are now always included as Build Inputs.
Previously they had to be specified in the .app.toml files as BuildInputs
manually.
Always adding them to BuildInputs is more reliable.
If those files are not listed as BuildInputs changing e.g. the build-command or
the upload destination won't trigger an necessary application rebuild.