docker: support unlimited path components in repository name
No slashes were supported in the repository name.
According to
https://github.com/docker/distribution/blob/master/docs/spec/api.md#overview
multiple path components, separated with a "/" are valid.
These are for example used in GCR.

The docker package is adapted to handle parse URIs with slashes in the
repository name correctly.

Support for specifying a host and port in the URI that is passed to
docker.Upload() is removed.
The functionality was not used, the server:port was always empty because it's
not supported to specify it in the .app.toml files.

This fixes: https://github.com/simplesurance/baur/issues/96