cfg: support to define multiple tasks in .app.toml files
The [Build] section in the .app.toml files is replaced with a [[Task]] section.
The section allows to define multiple tasks per application instead of only a
build-command.

For now only a single task named 'build' is supported.
In follow-up PRs baur will be refactored to support definition and running of
multiple tasks of an application.

Additional changes:
- replacing of variables in config files is moved to cfg/resolvers, variables
  can be resolved in the config struct via the Resolve() methods,
- merging config files with includes is now implemented in the cfg package,
  the package provides an IncludeDB object that caches already loaded includes,
  config structs can be merged with compatible Include struct via the Merge()
  methods,
- the cfg/app.go is splitted into multiple files per struct, it became too big
- $UUID variables in the config are now replaced by UUID v4 uuids, previously
  they were replaced with Mongo Object
  IDs(https://docs.mongodb.com/manual/reference/method/ObjectId/)
- printf %w is used instead of github.com/pkg/errors for error wrapping