introduce Task type
Task represents the section with the same name in the .app.cfg file.
The fields and method relating to Inputs, Outputs and building an App are moved
to Task.

In the future baur will support to define multiple tasks per app.
Currently only one task name with the name "build" is supported.