resolve: refactor resolvers
The configuration parameters for resolvers are moved from their structs to the
Resolve() methods.
This allows to use the same resolver object for resolving different paths.
The Resolver interface is removed. The Resolve functions do not have the same
signature anymore. Also the interface can be defined in the package that uses
the resolvers.