docker: credential were not found if registry was not a full URL
If the address of the registry used in the authentication config was not a
URL, baur did not find the credential entry.
Using hostnames or network addresses for the registry is common.

Do a simple string compassion when searching for the auth data for a registry
instead of assuming it's an URL.