improve error message when no repository config file is found
If no repository config file was found, commands failed with the message "file
not found".

Improve the error message, print which file was not found and where it is
expected.

To be able to print a more exact error message, the baur.FindRepository*
functions were replaced with functions that only return the path to the
repository config file instead of also instantiating a repository.