refactor evaluating the status of a task
A TaskStatusEvaluator struct is introduced that is responsible for evaluating
the status by resolving the build inputs, calculating the total input digest,
querying the storage for past task runs with the same input digest and then
returning a status.

The functionality is similiar to what baur.BuildStatus() did before + resolving
also the inputs.