make: pass -trimpath to go build
Pass -trimpath to the go compiler when building the baur binary:

	-trimpath
		remove all file system paths from the resulting executable.
		Instead of absolute file system paths, the recorded file names
		will begin with either "go" (for the standard library),
		or a module path@version (when using modules),
		or a plain import path (when using GOPATH).

The absolute path that are shown e.g. in panic messages to the source file are
in 99% cases only valid on the machine on that the binary was build.
On other machines the source files might not exist or reside in different paths.
The trimmed paths are better portable.