command: dynamically evaluate unit when printing sizes
Replace term.BytesToMib() with term.FormatSize()

Instead of always printing sizes in MiB commands will print it the size in
bytes if it is < 1KiB, in KiB if it's < 1 Mib, etc.

"ls outputs" uses term.FormatSize() only when "--csv" is not passed.
If "--csv" is passed, size is always in Bytes because it's easiest to parse and
process.