command: dynamically format duration
Instead of printing duration always in seconds, print them in bigger units when
the durations are big enough.

If they are >=1min they are printed with rounded second precision.
If they are >=1 Millisecond with rounded Millisecond precision.
Otherwise they are printed with the default Go String() formatter.

"ls outputs" prints durations as before always in seconds when "--csv" is
passed.