term: add option parameter to format in base unit without unit name suffix
FormatSize() and FormatDuration() now accepts an optional list of Options.
The only available option for now is FormatBaseWithoutUnitName(enabled).
When true is passed, the format functions return the value in it's base unit
without unit name suffix.

This is handy for formatting units differently when --csv is passed on the
commandline.