run: fix: recording run fails if baur-repository is not a git repository
When the baur repository was not a git repository, recording a task run failed
because the HEAD git commit could not be resolved.

How the repository state in the commands is refactored.
The git package is moved to vcs/git.
It got a new function that returns if a directory is part of a git repository.

A vcs package is introduced with a function vcs.GetState(path).
If path is in a git repository and the git tools are available in $PATH, it
returns a git.RepositoryState object.  Otherwise a dummy RepositoryState()
object, implementing the same interface, that always returns
vcs.ErrVCSRepositoryNotExist
vcs.GetState() caches the returned RepositoryState() for the dir that it was
called it.
The vcs package will also make it easier to add support for other VCS in the
future.

If vcs.ErrVCSRepositoryNotExist is returned by those methods when returning a
task run it is ignored and no vcs information are recorded.