cfg: validate include ID uniqueness before loading includes into the db
Validate when loading an include file that all IDs of includes in the file are
unique.
Previously an include ID only had to be unique per include type.
E.g. a include file could have a Taskinclude named "main" and an inputInclude
with the same name.
Enforce they are unique, it makes it easier to find issues when they are
referenced by their include ID in a filan include ID only had to be unique per
include type.
E.g. a include file could have a Taskinclude named "main" and an inputInclude
with the same name.
Enforce they are unique, it makes it easier to find issues when they are
referenced by their include ID in a file