cfg: fix: loading config failed when 2 configs references the same include file
When 2 app configs references the same include file, loading the later failed
because it was already loaded into the includedb.

When merging the task of the first include file, LoadOutputInclude() loaded the
include file from TaskMerge().
When merging the task of the second include file, LoadInputInclude() was called
for the same include id, it could find the reference in the db.inputs map,
tried to load it as input which then failed because it was already in the
db.outputs map

TODO: add testcase for it