resolver: refactor gosource testcases
The go project dirs for the testcases are now kept in testdata/.
Every testdata/ contains a test_config.json that describes the arguments that
are passed to the Gosource resolver and the expected files it should resolve to.
In the test_config.json the variable $WORKDIR can be used, it's replaced with
directory where the test_config.json resides.

The TestResolve() testcase runs a testcase for all subdirectories in testdata/.

This makes writing and maintaining testcases for the go-resolver much more
straightforward.