resolver: gosource: accept Queries instead of Paths
The gosource resolver now accepts a list of queries instead of app-directory
relative paths.
This allows more freedom when specifying the go files for which sources should
be resolved.

The queries are passed through to packages.Load().
It supports the same syntax then described at:
https://github.com/golang/tools/blob/bc8aaaa29e0665201b38fa5cb5d47826788fa249/go/packages/doc.go#L17.

- file=X to match files,
- query=X and queries with a '<something>=' prefix are passed through directly
  to the underlying build tool (go list normally)

Normally "go list" is used as build tool, which means the usual golang package
patterns can be used.

All queries are executed in the app directory.
All defines Queries in the slice are passed to pacakges.Load() at once.
Previously we called packages.Load() for each element in the Paths slice.

The cfg files are adapted accordingly.
When upgrading configs from v4, Paths are converted as:
"./<OLD-PATH>/...". This should ensure the same behavior then before.