resolver: gosource: support to query for go test files
Add new configration field called Tests for the gosource resolver.
If it's set to true, queries also resolve files to golang test files.
By default this is disabled.

This is useful when defining a baur task to run golang tests. With the option
it's possible to include the test files as inputs.