resolver: gosource: support fileglob queries
Support a fileglob query type.
When gosource resolver queries have the prefix "fileglob=", the glob pattern is
resolved to files. The doublestar glob is supported.
The resolved files are passed to the packages.Load() as filepaths.

This allows to e.g. resolve the sources of all **/*_test.go files in
go-unit-test tasks.