cfg: fix: app.Resolve() changes values of same included task instance
When variables of an included task were resolved, values changed in the same
slice instance.
The values changed for independent App instances were affected.

Fix it by deep copying the Task Include struct into a Task struct.

The deep copy is done via gob encoding/decoding. It's comparatively slow but
works reliable and was fast to implement.
Copying the values in the struct would have much better performance but is more
error prone. When fields changes in the cfg it must not be forgotten to adapt
the copy logic.
This could be solved by a testcase but then it becomes much more effort then
using gob.