cfg: convert the GolangSources field in the app config to a slice
The GolangSources field consisted of multiple field and app configs only had a
single GolangSources element.

When includes that contained GolangSources definitions were merged with the app
config, the values of the includes were appended to the single GolangSources
definition.
This could cause that golangsources were evaluated for a path with wrong
Environment, BuildFlags or Tests settings.

Convert the GolangSources to a slice.
This allows to keep the settings from GolangSources from includes and also
allows to specify multiple GolangSources definitions with different settings in
the .app.toml file.