simplify handling input Strings
Use the introduced Input interface to store Inputfiles and the inputString in
the same slice.
This simplifies the usage, conditional handling for the inputString can be
removed in many places.

InputString are now only added to Inputs if the string is not empty.
Previously InputString with an empty String value were added which made it
necessary to check later again if the String was empty or not.

In some places inputs.GetInputString().Exists() was called without checking if
GetInputString() returned nil.
This could have caused a nil pointer dereference. This is fixed by this change.

The commit also applies some minor cosmetic improvements.