Support to specify Strings that are hashed as Inputs via a commandline parameter (#185)
The run, status and ls inputs command support a new parameter called "--input-str" that allows to specify a string  on the commandline that is hashed as additional input.
The run and status command additionally have a new "--alt-input-str" parameter. It allows to specify an input string that is used for looking up past builds when no build could be found with the" --input-str" as input.