remove the InputsUndefined status
If a task has no inputs defined it is now treated as validation error.
There is no benefit of handling this case special and have InputsUndefined
status for it.

Because having no inputs defined is now a validation error, the app configs must
be merged with their includes before the validation is done.
Before it was possible to do the validation before the merge without getting
false errors.

Validating the upgraded configuration files is moved to the command
implementation. The validation is the normal process of loading all apps in the
repository.