test: reset environment variable envVarPSQLURL when testcase finishes
The testcase sets the environment variable envVarPSQLURL to a specific database
url.

Before the next testcase runs, the environment variable must be unset again.
If this is not done, the environment variable is still set when the next
testcase runs, the wrong database might be used which can cause test failures.